refactor: extract css generation logic to modules
